# hospital-service/docs/swagger.yaml
basePath: /
definitions:
...
  doctor.Doctor:
    properties:
      created_at:
        type: string
      doctor_id:
        type: string
      email:
        type: string
      full_name:
        type: string
      hospital_id:
        type: string
      image:
        type: string
      phone:
        type: string
      slug:
        type: string
      specialty:
        $ref: '#/definitions/enums.Specialty'
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
  enums.Specialty:
    enum:
    - ophthalmology
    - retina
    - cornea
    - glaucoma
    - refractive_surgery
    - pediatric_ophthalm
    - schedule_manager
    type: string
    x-enum-comments:
      SpecialtyCornea: Chuyên khoa giác mạc
      SpecialtyGlaucoma: Cườm nước
      SpecialtyOphthalmology: Nhãn khoa tổng quát
      SpecialtyPediatricOphthalm: Nhãn khoa nhi
      SpecialtyRefractiveSurgery: Phẫu thuật khúc xạ (LASIK)
      SpecialtyRetina: Chuyên khoa võng mạc
      SpecialtyScheduleManager: Quản lý lịch bác sĩ
    x-enum-descriptions:
    - Nhãn khoa tổng quát
    - Chuyên khoa võng mạc
    - Chuyên khoa giác mạc
    - Cườm nước
    - Phẫu thuật khúc xạ (LASIK)
    - Nhãn khoa nhi
    - Quản lý lịch bác sĩ
    x-enum-varnames:
    - SpecialtyOphthalmology
    - SpecialtyRetina
    - SpecialtyCornea
    - SpecialtyGlaucoma
    - SpecialtyRefractiveSurgery
    - SpecialtyPediatricOphthalm
    - SpecialtyScheduleManager
  hospital.Hospital:
    properties:
      address:
        type: string
      city:
        type: string
      created_at:
        type: string
      doctors:
        items:
          $ref: '#/definitions/doctor.Doctor'
        type: array
      email:
        type: string
      hospital_id:
        type: string
      image:
        type: string
      latitude:
        type: number
      longitude:
        type: number
      name:
        type: string
      phone:
        type: string
      slug:
        type: string
      updated_at:
        type: string
      url_map:
        type: string
      ward:
        type: string
    type: object
  hospitalhandler.nearbyHospitalReq:
    properties:
      latitude:
        type: number
      longitude:
        type: number
      radius_km:
        type: number
    required:
    - latitude
    - longitude
    - radius_km
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
host: localhost:8084
info:
  contact: {}
  title: Auth Service API
  version: "1.0"
paths:
...
  /hospitals:
    get:
      description: Retrieve all hospitals with optional filters
      parameters:
      - description: Filter by hospital name (partial match)
        in: query
        name: name
        type: string
      - description: Filter by address (partial match)
        in: query
        name: address
        type: string
      - description: Filter by city (exact match)
        in: query
        name: city
        type: string
      - description: Filter by ward (exact match)
        in: query
        name: ward
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hospital.Hospital'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all hospitals
      tags:
      - Hospitals
    post:
      consumes:
      - multipart/form-data
      description: Add hospital info with optional logo upload
      parameters:
      - description: Hospital Name
        in: formData
        name: name
        required: true
        type: string
      - description: Address
        in: formData
        name: address
        type: string
      - description: Phone
        in: formData
        name: phone
        type: string
      - description: Email
        in: formData
        name: email
        type: string
      - description: Url Map
        in: formData
        name: url_map
        type: string
      - description: Ward
        in: formData
        name: ward
        type: string
      - description: City
        in: formData
        name: city
        type: string
      - description: Hospital Logo
        in: formData
        name: logo
        type: file
      - description: Latitude
        in: formData
        name: latitude
        required: true
        type: number
      - description: Longitude
        in: formData
        name: longitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Create a new hospital
      tags:
      - Hospitals
  /hospitals/{hospital_id}:
    delete:
      description: Delete hospital by hospital ID
      parameters:
      - description: Hospital ID
        in: path
        name: hospital_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete hospital
      tags:
      - Hospitals
    get:
      description: Retrieve hospital record using hospital ID
      parameters:
      - description: Hospital ID
        in: path
        name: hospital_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/hospital.Hospital'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get hospital by ID
      tags:
      - Hospitals
    put:
      consumes:
      - multipart/form-data
      description: Update hospital info and logo
      parameters:
      - description: Hospital ID
        in: path
        name: hospital_id
        required: true
        type: string
      - description: Hospital Name
        in: formData
        name: name
        type: string
      - description: Address
        in: formData
        name: address
        type: string
      - description: Phone
        in: formData
        name: phone
        type: string
      - description: Email
        in: formData
        name: email
        type: string
      - description: Url Map
        in: formData
        name: url_map
        type: string
      - description: Ward
        in: formData
        name: ward
        type: string
      - description: City
        in: formData
        name: city
        type: string
      - description: Hospital Logo
        in: formData
        name: logo
        type: file
      - description: Latitude
        in: formData
        name: latitude
        type: number
      - description: Longitude
        in: formData
        name: longitude
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Update a hospital
      tags:
      - Hospitals
  /hospitals/cities:
    get:
      description: Retrieve all distinct cities that have hospitals
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: List all cities
      tags:
      - Hospitals
  /hospitals/filter:
    get:
      description: Retrieve hospitals filtered by city and ward (both optional, can
        filter by either or both)
      parameters:
      - description: City name
        in: query
        name: city
        type: string
      - description: Ward name
        in: query
        name: ward
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hospital.Hospital'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: List hospitals by city and ward
      tags:
      - Hospitals
  /hospitals/nearby:
    post:
      consumes:
      - application/json
      description: Find hospitals within a radius (km) from given coordinates
      parameters:
      - description: Latitude, Longitude and Radius (km)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/hospitalhandler.nearbyHospitalReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hospital.Hospital'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Find nearby hospitals
      tags:
      - Hospitals
  /hospitals/search:
    get:
      description: Search hospitals by keyword in name, address, ward, or city
      parameters:
      - description: Search keyword
        in: query
        name: keyword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/hospital.Hospital'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Search hospitals by address
      tags:
      - Hospitals
  /hospitals/slug/{slug}:
    get:
      description: Retrieve hospital record using hospital slug
      parameters:
      - description: Hospital Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get hospital by slug
      tags:
      - Hospitals
  /hospitals/wards:
    get:
      description: Retrieve all distinct wards for a given city
      parameters:
      - description: City name
        in: query
        name: city
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: List wards by city
      tags:
      - Hospitals
...
swagger: "2.0"